Clean up ReferenceListerService
